Move comparator and parse request filter.
